+++ src/main/java/subway/controller/section/dto/SectionCreateRequest.java
@@ -1,0 +1,29 @@
+package subway.controller.section.dto;
+
+public class SectionCreateRequest {
+    private String upStationName;
+    private String downStationName;
+    private int distance;
+
+    public SectionCreateRequest(String upStationName, String downStationName, int distance) {
+        this.upStationName = upStationName;
+        this.downStationName = downStationName;
+        this.distance = distance;
+    }
+
+    private SectionCreateRequest() {
+
+    }
+
+    public String getUpStationName() {
+        return upStationName;
+    }
+
+    public String getDownStationName() {
+        return downStationName;
+    }
+
+    public int getDistance() {
+        return distance;
+    }
+}
