+++ src/main/java/subway/controller/station/dto/StationCreateRequest.java
@@ -1,0 +1,16 @@
+package subway.controller.station.dto;
+
+public class StationCreateRequest {
+    private String name;
+
+    public StationCreateRequest() {
+    }
+
+    public StationCreateRequest(String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
