--- src/main/java/subway/domain/Section.java
@@ -1,54 +1,0 @@
-package subway.domain;
-
-public class Section {
-    private final Long id;
-    private final Station upStation;
-    private final Station downStation;
-    private final Distance distance;
-
-    public Section(Long id, Station upStation, Station downStation, Distance distance) {
-        this.id = id;
-        this.upStation = upStation;
-        this.downStation = downStation;
-        this.distance = distance;
-    }
-
-    public Section(Station upStation, Station downStation, Distance distance) {
-        this(null, upStation, downStation, distance);
-    }
-
-    public boolean contains(Station station) {
-        return upStation.equals(station) || downStation.equals(station);
-    }
-
-    public Distance calculateNewSectionDistance(Distance another) {
-        return this.distance.reduce(another);
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public Station getUpStation() {
-        return upStation;
-    }
-
-    public Station getDownStation() {
-        return downStation;
-    }
-
-    public Distance getDistance() {
-        return distance;
-    }
-
-
-    @Override
-    public String toString() {
-        return "Section{" +
-                "id=" + id +
-                ", upStation=" + upStation +
-                ", downStation=" + downStation +
-                ", distance=" + distance +
-                '}';
-    }
-}
