--- src/main/java/subway/domain/Station.java
@@ -1,53 +1,0 @@
-package subway.domain;
-
-import java.util.Objects;
-
-public class Station {
-    private Long id;
-    private String name;
-
-    public Station() {
-    }
-
-    public Station(Long id, String name) {
-        this.id = id;
-        this.name = name;
-    }
-
-    public Station(String name) {
-        this.name = name;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public boolean isSame(Station anotherStation) {
-        return this.name.equals(anotherStation.name);
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Station station = (Station) o;
-        return id.equals(station.id) && name.equals(station.name);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id, name);
-    }
-
-    @Override
-    public String toString() {
-        return "Station{" +
-                "id=" + id +
-                ", name='" + name + '\'' +
-                '}';
-    }
-}
