+++ src/main/java/subway/persistence/dao/LineDao.java
@@ -6,7 +6,7 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import subway.domain.Line;
+import subway.service.line.domain.Line;
 
 import javax.sql.DataSource;
 import java.util.List;
