+++ src/main/java/subway/persistence/dao/StationDao.java
@@ -9,8 +9,8 @@ import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import subway.domain.Station;
 import subway.persistence.dao.entity.StationEntity;
+import subway.service.station.domain.Station;
 
 import javax.sql.DataSource;
 import java.util.List;
