+++ src/main/java/subway/service/line/LineService.java
@@ -1,0 +1,53 @@
+package subway.service.line;
+
+import org.springframework.stereotype.Service;
+import subway.persistence.dao.LineDao;
+import subway.service.line.domain.Line;
+import subway.service.line.dto.LineRequest;
+import subway.service.line.dto.LineResponse;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+@Service
+public class LineService {
+    private final LineDao lineDao;
+
+    public LineService(LineDao lineDao) {
+        this.lineDao = lineDao;
+    }
+
+    public LineResponse saveLine(LineRequest request) {
+        Line persistLine = lineDao.insert(new Line(request.getName(), request.getColor()));
+        return LineResponse.of(persistLine);
+    }
+
+    public List<LineResponse> findLineResponses() {
+        List<Line> persistLines = findLines();
+        return persistLines.stream()
+                .map(LineResponse::of)
+                .collect(Collectors.toList());
+    }
+
+    public List<Line> findLines() {
+        return lineDao.findAll();
+    }
+
+    public LineResponse findLineResponseById(Long id) {
+        Line persistLine = findLineById(id);
+        return LineResponse.of(persistLine);
+    }
+
+    public Line findLineById(Long id) {
+        return lineDao.findById(id);
+    }
+
+    public void updateLine(Long id, LineRequest lineUpdateRequest) {
+        lineDao.update(new Line(id, lineUpdateRequest.getName(), lineUpdateRequest.getColor()));
+    }
+
+    public void deleteLineById(Long id) {
+        lineDao.deleteById(id);
+    }
+
+}
