+++ src/main/java/subway/service/line/dto/LineRequest.java
@@ -1,0 +1,23 @@
+package subway.service.line.dto;
+
+public class LineRequest {
+    private String name;
+    private String color;
+
+    public LineRequest() {
+    }
+
+    public LineRequest(String name, String color) {
+        this.name = name;
+        this.color = color;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getColor() {
+        return color;
+    }
+
+}
