+++ src/main/java/subway/service/line/dto/LineResponse.java
@@ -1,0 +1,31 @@
+package subway.service.line.dto;
+
+import subway.service.line.domain.Line;
+
+public class LineResponse {
+    private final Long id;
+    private final String name;
+    private final String color;
+
+    public LineResponse(Long id, String name, String color) {
+        this.id = id;
+        this.name = name;
+        this.color = color;
+    }
+
+    public static LineResponse of(Line line) {
+        return new LineResponse(line.getId(), line.getName(), line.getColor());
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getColor() {
+        return color;
+    }
+}
