--- src/main/java/subway/ui/StationController.java
@@ -1,57 +1,0 @@
-package subway.ui;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-import subway.application.station.StationService;
-import subway.application.station.dto.StationCreateResponse;
-import subway.dto.StationRequest;
-import subway.dto.StationResponse;
-import subway.ui.dto.StationCreateRequest;
-
-import java.net.URI;
-import java.sql.SQLException;
-import java.util.List;
-
-@RestController
-@RequestMapping("/stations")
-public class StationController {
-    private final StationService stationService;
-
-    public StationController(StationService stationService) {
-        this.stationService = stationService;
-    }
-
-    @PostMapping
-    public ResponseEntity<StationCreateResponse> createStation(@RequestBody StationRequest stationRequest) {
-        StationCreateRequest stationCreateRequest = new StationCreateRequest(stationRequest.getName());
-        StationCreateResponse station = stationService.saveStation(stationCreateRequest);
-        return ResponseEntity.created(URI.create("/stations/" + station.getId())).body(station);
-    }
-
-    @GetMapping
-    public ResponseEntity<List<StationResponse>> showStations() {
-        return ResponseEntity.ok().body(stationService.findAllStationResponses());
-    }
-
-    @GetMapping("/{id}")
-    public ResponseEntity<StationCreateResponse> showStation(@PathVariable Long id) {
-        return ResponseEntity.ok().body(stationService.findStationById(id));
-    }
-
-    @PutMapping("/{id}")
-    public ResponseEntity<Void> updateStation(@PathVariable Long id, @RequestBody StationRequest stationRequest) {
-        stationService.updateStation(id, stationRequest);
-        return ResponseEntity.ok().build();
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteStation(@PathVariable Long id) {
-        stationService.deleteStationById(id);
-        return ResponseEntity.noContent().build();
-    }
-
-    @ExceptionHandler(SQLException.class)
-    public ResponseEntity<Void> handleSQLException() {
-        return ResponseEntity.badRequest().build();
-    }
-}
