--- src/main/java/subway/ui/dto/StationCreateRequest.java
@@ -1,16 +1,0 @@
-package subway.ui.dto;
-
-public class StationCreateRequest {
-    private String name;
-
-    public StationCreateRequest() {
-    }
-
-    public StationCreateRequest(String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
