--- src/main/java/subway/ui/section/dto/SectionCreateRequest.java
@@ -1,29 +1,0 @@
-package subway.ui.section.dto;
-
-public class SectionCreateRequest {
-    private String upStationName;
-    private String downStationName;
-    private int distance;
-
-    public SectionCreateRequest(String upStationName, String downStationName, int distance) {
-        this.upStationName = upStationName;
-        this.downStationName = downStationName;
-        this.distance = distance;
-    }
-
-    private SectionCreateRequest() {
-
-    }
-
-    public String getUpStationName() {
-        return upStationName;
-    }
-
-    public String getDownStationName() {
-        return downStationName;
-    }
-
-    public int getDistance() {
-        return distance;
-    }
-}
