+++ src/test/java/subway/application/section/SectionServiceTest.java
@@ -4,6 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import subway.persistence.dao.SectionDao;
+import subway.service.section.SectionService;
 
 @SuppressWarnings("NonAsciiCharacters")
 @SpringBootTest
