+++ src/test/java/subway/dao/StationDaoTest.java
@@ -5,8 +5,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
-import subway.domain.Station;
 import subway.persistence.dao.StationDao;
+import subway.service.station.domain.Station;
 
 import javax.sql.DataSource;
